New .hex file with clock source from internal RC
The only change in the DFU firmware was made at the source of the clock for the SoftDevice application. With the original, custom boards without the 32.768 crystal could not use the pre-compiled Sparfun bootloader available.

The "nRF52832 Breakout Board Hookup Guide" induces to understand that it could be without:

"The nRF52832 Breakout equips the chip with a 32.768kHz crystal -- connected to GPIO pins 0 and 1. Unfortunately, those pins make up a quarter of the available ADC inputs. So, if your application doesn't require an RTC -- and you need those pins to for another purpose -- grab a hobby knife and cut the two traces between their pads to disconnect the crystal."